Move ContentRangeParser to datasources.utils, simplify exception handling
Consolidate duplicate catch clauses in GCS storage object using
instanceof pattern matching with a single throw per error path.

Developed using AI-assisted tooling
