+++ x-pack/plugin/esql/src/test/java/org/elasticsearch/xpack/esql/datasources/utils/ContentRangeParserTests.java
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-package org.elasticsearch.xpack.esql.datasources;
+package org.elasticsearch.xpack.esql.datasources.utils;
 
 import org.elasticsearch.test.ESTestCase;
